Refactor VirtualAudioContext  (#2)
* reuse patch code for creating new nodes

The `createNode` function duplicated some code from where we were
applying patches, which resulted in some odd discrepancies.

While refactoring, I also added a check to make sure that if we are
trying to set the value of an AudioParam, it really _is_ an AudioParam.

* merge some convergent logic

* unlink AudioParam Properties from ScheduledUpdates

The `AudioParam` variant of `Property` is used to set an AudioParam to a
fixed value, but a 'ScheduledUpdate' is asking it to change over time
(or just _at_ a certain time). These two things don't really overlap, so
it's strange that one should depend on the other, e.g., Why would we
want to wait until the scheduled updates are done before deleting a
fixed value we gave earlier? And why would we want to cancel our
scheduled value when we remove the fixed one?